+++ src/config/configStore.ts
@@ -77,7 +77,8 @@ export abstract class BaseConfigStore<
     P extends ConfigContents = ConfigContents
   >
   extends AsyncOptionalCreatable<T>
-  implements ConfigStore<P> {
+  implements ConfigStore<P>
+{
   protected static encryptedKeys: Array<string | RegExp> = [];
 
   protected options: T;
@@ -307,6 +308,7 @@ export abstract class BaseConfigStore<
    *
    * @param obj The object.
    */
+  // eslint-disable-next-line @typescript-eslint/ban-types
   public setContentsFromObject<U extends object>(obj: U): void {
     this.contents = {} as P;
     Object.entries(obj).forEach(([key, value]) => {