+++ src/org/permissionSetAssignment.ts
@@ -125,9 +125,10 @@ export class PermissionSetAssignment {
    *
    * @param permSetString The permission set string.
    */
-  private parsePermissionSetString(
-    permSetString: string
-  ): { nsPrefix: Optional<string>; permSetName: Optional<string> } {
+  private parsePermissionSetString(permSetString: string): {
+    nsPrefix: Optional<string>;
+    permSetName: Optional<string>;
+  } {
     const nsPrefixMatch = RegExp(/(\w+(?=__))(__)(.*)/).exec(permSetString);
 
     let nsPrefix: Optional<string>;