+++ src/testSetup.ts
@@ -4,7 +4,7 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-
+/* eslint-disable @typescript-eslint/ban-ts-comment */
 import { randomBytes } from 'crypto';
 import { EventEmitter } from 'events';
 import { tmpdir as osTmpdir } from 'os';
@@ -340,7 +340,7 @@ export const instantiateContext = (sinon?: any) => {
 export const stubContext = (testContext: TestContext) => {
   // Turn off the interoperability feature so that we don't have to mock
   // the old .sfdx config files
-  // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore
   Global.SFDX_INTEROPERABILITY = false;
 
@@ -363,8 +363,6 @@ export const stubContext = (testContext: TestContext) => {
     // Since read is now stubbed, make sure to call getPath to initialize it.
     configFile.getPath();
 
-    // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-    // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
     // @ts-ignore set this to true to avoid an infinite loop in tests when reading config files.
     configFile.hasRead = true;
     return stub;
@@ -391,7 +389,6 @@ export const stubContext = (testContext: TestContext) => {
   };
 
   // Mock out all config file IO for all tests. They can restore individually if they need original functionality.
-  // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
   // @ts-ignore
   testContext.SANDBOXES.CONFIG.stub(ConfigFile.prototype, 'readSync').callsFake(readSync);
   testContext.SANDBOXES.CONFIG.stub(ConfigFile.prototype, 'read').callsFake(read);
@@ -432,7 +429,8 @@ export const stubContext = (testContext: TestContext) => {
   stubMethod(testContext.SANDBOXES.CRYPTO, Crypto.prototype, 'getKeyChain').callsFake(() =>
     Promise.resolve({
       setPassword: () => Promise.resolve(),
-      getPassword: (data: object, cb: AnyFunction) => cb(undefined, '12345678901234567890123456789012'),
+      getPassword: (data: Record<string, unknown>, cb: AnyFunction) =>
+        cb(undefined, '12345678901234567890123456789012'),
     })
   );
 
@@ -486,7 +484,6 @@ const _testSetup = (sinon?: any) => {
 
   beforeEach(() => {
     // Allow each test to have their own config aggregator
-    // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
     // @ts-ignore clear for testing.
     delete ConfigAggregator.instance;